Fill-out emitter with missing node emitter methods
Added the methods still missing from this emitter that are in the native node emitter.  These are eventNames(), listenerCount() and removeAllListeners().  Have added some overloading so if you call listenerCount() or removeAllListeners() without an event name it operates on all events.

Also, added broadcast and broadcastAsync() to emitter from global emitter.