Add some inline documentation, fix namespacing error and make more efficient
Namespacing should be against a name or symbol, not an object.  Fixed this so name are reference to unique objects that can be used against a WeakMap.  Also, ensured the TargetEventEmitter is actually a singleton for given namespace and added factory method to ensure no unnecessary construction.

Added inline documentation (not complete yet) and change name from GlobalEventEmitter to TargetEventEmitter.  This is a better name but still not great.  It might be renamed again.