Bump to version 0.2.0
We have the basic emitter now that does everything (and a bit more) that an emitter with emit targets and parent/child relationship should do.  The focus of the next cycle needs to be on solidness or robustness in the code.  This could mean reducing code or removing some methods so API still not stable.